Update to autotools used by binutils
When cgen source is installed in the top source directory of
binutils-gdb and configuring with --enable-maintainer-mode
--enable-cgen-maint=yes it is somewhat annoying to regenerate cgen
files without differences due to using different autotools.
Particularly so since the generated files in the git repository don't
all use the same autotools.

This patch fixes that by modernising the source a little and
regenerating with autoconf-2.69 and automake-1.15.1.

Note that doc/stamp-vti and doc/version.texi contain dates taken from
your checked out doc/cgen.texi file time stamp.  So in order to
regenerate these files consistently it is necessary to
  touch --date=xxx doc/cgen.texi
with the date taken from the last doc/cgen.texi commit.

	* configure.ac: Renamed from configure.in.
	(AC_INIT, AM_INIT_AUTOMAKE): Modernize.
	(AC_CONFIG_SRCDIR): Invoke.
	* INSTALL: Adjust for configure.ac renaming.
	* Makefile.am (AUTOMAKE_OPTIONS): Delete cygnus and version.
	Add foreign.
	* testsuite/Makefile.am (AUTOMAKE_OPTIONS): Likewise.
	* doc/Makefile.am (AUTOMAKE_OPTIONS): Delete cygnus, add foreign,
	info-in-builddir and no-texinfo.tex.
	(DISTCLEANFILES): Delete automake-1.9 hack.
	* Makefile.in: Regenerate.
	* aclocal.m4: Regenerate.
	* configure: Regenerate.
	* doc/Makefile.in: Regenerate.
	* doc/stamp-vti: Regenerate.
	* testsuite/Makefile.in: Regenerate.