rtx-funcs.scm: don't try to let-bind macros
As we look to move to guile 2+ one issue that we run into is the
difference in how macro expansion works between guile 1 and guile 2+.

In this particular case we create a macro, e.g. define-rtx-node, then
we say:

  (define (def-rtx-funcs)
    (let ((drn define-rtx-node))
      (drn .....)
      (drn .....)))

The problem is that guile 2+ can't handle the let binding of the
macro.

After transitioning to guile 2+ we could potentially look at the
possibility of using let-syntax here instead.  However, for now I
don't see this as essential.  Lets just remove the let binding, and
expand out the `drn` to `define-rtx-node`.

In the real code there's actually 4 different macros being bound.

There is no functional change after this commit.

ChangeLog:

	* rtx-funcs.scm (def-rtx-funcs): Don't try to let-bind macros,
	instead use long form macro names throughout.