pgmr-tools.scm: move local defines before function body
As we look to move to guile 2+ one issue we run into is code like
this:

  (define (some-func arg)
    (some-other-func ....)
    (define (helper-func arg)
       .....)
    (helper-func ....))

The problem here is `helper-func` defined within the body of
`some-func`.  This is fine for guile 1, but not allowed in guile 2+.
Instead `helper-func` should be defined before the body of
`some-func`.

It is fine to just reorder things, like:

  (define (some-func arg)
    (define (helper-func arg)
       .....)
    (some-other-func ....)
    (helper-func ....))

But to make it even clearer I have added a (begin ...), like this:

  (define (some-func arg)
    (define (helper-func arg)
       .....)
    (begin
      (some-other-func ....)
      (helper-func ....)))

There is no functional difference for this code on guile 1, but this
code will now compile fine with guile 2+.

ChangeLog:

	* pgmr-tools.scm (pgmr-lookup-insn): Ensure local defines occur
	before the function body.