cos.scm: avoid nested defines for creating top-level symbols
As we look to move to guile 2+ code like this is a problem:

  (if (... condition ....)
    (define (symbol args) body1)
    (define (symbol args) body2))

Guile 2+ will give an error like:

  definition in expression context, where definitions are not allowed, in form .....

In the particular case addressed in this commit we are creating an
alias for `vector-copy', but the code is handling the case where there
is no vector-copy available.

There is a `vector-copy' in guile 1.8.8 and later.  There is no
`vector-copy' in guile 1.6.8.  However, CGEN already doesn't work on
guile 1.6.8 for other reasons, so I think adding another restriction
is fine.

There are also hints that CGEN can be run using alternative scheme
environments (i.e. not guile).  Long term I don't see this as
something we would want to continue to support, but for now, if anyone
was doing such a thing, and their scheme doesn't have `vector-copy',
then they can add this:

  (define (vector-copy v)
    (list->vector (vector->list v)))

to their equivalent of guile.scm.

ChangeLog:

	* cos.scm (/object-vector-copy): Convert to a single top-level
	define.