cos.scm: move helper functions inside the macro that uses them
As we look to move to guile 2+ we need to consider how macro handling
has changed in guile.

In guile 2+ macros are expanded at compilation time, at which point
any helper functions they need must be available.  Thus code like
this is no longer valid:

  (define (some-func ....)
    ....)

  (defmacro some-macro (.....)
    (let ((local (some-func ....)))
      ....))

Will no longer work as `some-func' is not available for use at the
point when the macro `some-macro' is expanded.

There are two easy fixes for this code, one is to use eval-when to
evaluate the define of `some-func' twice, once at expand time, and
then again at eval time.  The other solution is to move the definition
of `some-func' inside `some-macro'.

The benefit of the second solution is that it works with older
versions of guile and so we can make the change now.  It also feels
like the cleaner solution if `some-func' is only used within
`some-macro', which is the case in the situation being addressed in
this patch.

ChangeLog:

	* cos.scm (/object-count-true): Move inside define-class macro.
	(/parse-member-list): Likewise.
	(/build-getter-defs): Likewise.
	(/build-setter-defs): Likewise.