Introduce eval-when macro and wrap nil in it
As we look to move to guile 2+ we need to adapt to guile's multi-phase
execution process.  Macros are now expanded during a separate expand
and compile phase, while general defines are evaluated at a later
evaluation phase.

As a result it is not possible to use the result of a general define
within a macro.

If we need to do this then the solution (for guile 2+) is to wrap the
global define in an `eval-when'.  Consider this example:

  (define magic-value 5)

  (defmacro blah (name arg)
    (if (equal? arg magic-value)
      (error "invalid argument value"))
    `(define ,name ',arg))

  (blah foo 2)
  (blah bar 5)

This works fine with guile 1.8.x, but gives an unbound variable
warning for `magic-value' with guile 2+.

If we wrap the definition of `magic-value' like this:

  (eval-when (expand load eval)
    (define magic-value 5))

Then the problem resolves itself as `magic-value' is now defined both
at macro expansion time, and at evaluation time.

ChangeLog:

	* guile.scm (eval-when): New macro.
	* utils.scm (nil): Wrap in `eval-when' macro.