unconditionally define eval1 function
The eval1 function is define using the form:

  (if some-condition
    (define (eval1 ...)
       ...)
    (define (eval1 ...)
       ...)

As we look to move to guile 2+ such conditional defines are not
allowed and result in an error like:

  definition in expression context, where definitions are not allowed, in form ....

The function eval1 is only defined like this in order to support
versions of guile before 1.6.

Testing shows that CGEN currently does not run successfully with even
version 1.6 of guile.

As such, I propose that everything before guile 1.8 be declared
unsupported.

Given that, we can simplify the definition of eval1, making it
unconditional.

ChangeLog:

	* guile.scm (eval1): Unconditionally define as we no longer
	support versions of guile before 1.8.