Add cgen.sh wrapper to invoke CGEN
Currently CGEN runs using Guile.  In theory we could imagine a
situation where CGEN is written in standard scheme and could be run on
any scheme implementation, but I don't think we're there yet.

Until recently CGEN would only run using Guile 1.8.x.  Recent commits
have changed this situation, and CGEN will now run using 2.x and 3.x.
With one massive caveat.

First, a disclaimer:  I don't claim to be either a scheme expert, not
a Guile expert.  What follows is simply my impressions and
observations from the work I've done over the last few months.  I
would love to be shown just how wrong I am with what I think the
situation is, but until then, what follows, is my best understanding
of the situation.

With Guile 2.x, Guile gained an auto-compilation system.  However,
this system is fundamentally broken and must be disabled in order to
provide a reliable user experience.

The reason for this claim is that Guile does not track dependencies
between auto-compiled files, thus, if file A includes file B, and the
compilation of file A depends on the contents of file B, then A will
not be recompiled even if B changes.

As a more concrete example:

  -- b.scm --

  (define-macro (foo)
    '(begin
       (display "Hello World")
       (newline)))

  -- a.scm --

  (include "b.scm")
  (foo)

  -- END --

Run this as:

  $ guile -s a.scm
  Hello World

Now modify b.scm, changing 'Hello World' to 'Goodbye All', then rerun:

  $ guile -s a.scm
  Hello World

The change is b.scm is never seen.

This is obviously a problem during CGEN development, if I modify a
macro then the modified macro will not be picked up.  But also, given
the way CGEN is generally used (I think), that is, straight from the
development repository, every CGEN user is basically a CGEN developer;
if they use git to fetch the latest CGEN changes then a change might
land only in b.scm, updating the macro.  If the CGEN user doesn't know
they need to force a recompile then they will fail to get the benefits
of the latest changes.

I've asked about this issue on the guile-users list, and the answer
appears to be that Guile has morphed from being an interpreted
language into some sort of hybrid.  The expectation is that we would
now follow a compile and install approach, however, we would still
need to install both the .scm files and the compiled .go files.
This (to me) feels like a terrible idea as an end user might still
think they can modify the .scm file and see the results, when, as
detailed above, this is not reliably the case.

Long term for CGEN we might need to consider moving to a different
scheme implementation, or maybe porting CGEN to a different language
entirely.

However, in the short term I propose this patch.

Currently, CGEN users invoke Guile directly, but I propose that we add
a new wrapper script (cgen.sh) within the CGEN repository.  This
script will take on the task of invoking Guile, and will ensure that
Guile's broken auto-compilation is turned off.

Even with this script in place then CGEN can still be run the old
way, with guile 1.8.8 this works just as it did before, but with guile
2+ the old way works fine, so long as Guile's auto compilation is
turned off (either by adding extra flags to the guile command line, or
setting the appropriate environment variable).

ChangeLog:

	* cgen.sh: New file.