remove symbol-bound? definition, and all remaining uses
The function `symbol-bound?` was defined conditionally.  As we move to
guile 2+ such conditional defines are not allowed, defines must appear
at the top level.

As CGEN now only supports guile 1.8+ we can simplify the definition of
`symbol-bound?` by removing support for older versions of guile.

However, when I look through the uses of `symbol-bound?` I notice that
all of the uses can similarly be simplified, instead of checking to
see if the symbol is defined, lets just assume it is (or isn't as
appropriate) based on the knowledge that we will not be using a very
old version of guile.

In guile.scm, I now unconditionally redefine `load`.  This conditional
definition was broken anyway, we had this:

    (if (symbol-bound? #f 'load-from-path)
        (begin
          (define (load file)
           (begin
             ;(load-from-path file)
             (primitive-load-path file)
             ))
          )

Clearly at one point we were using `load-from-path`, but when this
changed the symbol-bound? check wasn't updated.  Given that both
primitive-load-path and load-from-path exist in guile 1.8+, I now just
unconditionally redefine load.

I unconditionally define `%stat`.  This doesn't exist in guile 1.8+.

The `reverse!` function does exist in guile 1.8+, so the conditional
definition of this function can now be deleted.

Finally (in guile.scm), debug-enable exists in guile 1.8+, so the call
to this function doesn't need to be conditional.

In utils.scm I deleted some old, commented out, code relating to
symbol-bound?, then in `bound-symbol?` I use `module-defined`
directly.  This is what was used inside the old body of
`symbol-bound?`.

ChangeLog:

	* guile.scm (symbol-bound?): Delete.
	(load): Always redefine.
	(%stat): Likewise.
	(reverse!): Delete.
	(debug-enable 'backtrace): Call unconditionally.
	* utils.scm: Delete commented out code relating to symbol-bound?.
	(bound-symbol?): Use module-defined directly.