handle debug features that have been removed in guile 2+
As we look to move CGEN to guile 2+ one issue is that later versions
of guile support a smaller set of debug features (for use with
debug-enable and debug-set!) than guile 1.8.

This commit splits the debug features into the unconditional, and
those that only exist on the earlier versions of guile.

There should be no change after this commit when using guile 1.8.

ChangeLog:

	* guile.scm: Don't try to set debug features that have been
	removed in guile 2+.