Stop using defmacro:syntax-transformer
As we look to support guile 2+, defmacro:syntax-transformer is no
longer supported.  The comment attached with this code indicates it
was required to support 'Hobbit', I guess this was some alternative
scheme, but I can find no trace of it online.

Given that CGEN only really supports guile for now, I propose that we
simple remove the use of defmacro:syntax-transformer and instead just
create a macro that will be supported for guile 1.8.x and guile 2+.

ChangeLog:

	* rtl.scm (define-rtx-node): Rewrite as a macro without the use of
	defmacro:syntax-transformer.
	(define-rtx-syntax-node): Likewise.
	(define-rtx-operand-node): Likewise.
	(define-rtx-macro-node): Likewise.