Remove use of procedure->memoizing-macro
As we look to move to guile 2+, procedure->memoizing-macro is no
longer available.  This procedure is only used in two places in order
to try an optimise object field get/set calls.

The optimisation tries to speed up calls when the field being get/set
is on the 'self object, in which case the validity check for the
field's existence can be done at macro expansion time.

Rather than try to preserve the optimisation, in this commit I propose
to just remove it entirely.  All field get/set will take the slow
path, that is first validate the field exists, then access the field.
This is the easiest solution and allows us to move towards guile 2+
support.

We can always look at optimising this again in the future if there's
felt to be a need.

ChangeLog:

	* cos.scm (/elm-make-method-getter): Delete.
	(elm-get): Remove use of /elm-make-method-getter.
	(/elm-make-method-setter): Delete.
	(elm-set!): Remove use of /elm-make-method-setter.