Don't try to use logit until it has been defined
In read.scm the /cmd-define-rtl-version function makes use of the
logit macro before the macro has been defined.  This is fine in older
versions of guile as the macro would not be expanded until the
function was executed, but as we move to newer versions of guile where
macros can be expanded earlier (as the function is first parsed) then
we need to ensure the macros are defined when they are used.

This commit moves the definition of /cmd-define-rtl-version later in
the file, after the point where utils.scm (in which logit is defined)
is included.

ChangeLog:

	* read.scm (/cmd-define-rtl-version): Move to after utils.scm has
	been loaded.