prevent stack overflow with cgen-opc for m32c with guile 2.0.14
When generating the opcodes file (using tool cgen-opc.scm) for the
m32c target using guile 2.0.14 I would run into stack overflow in the
function gen-enum-decl.

The core of this function builds a list of strings, which are then
merged into a single string with string-list->string.

The m32c target ends up with 84,204 strings in the string list being
generated at one point, and, on guile 2.0.14, this seems to result in
stack overflow.

The same stack overflow doesn't occur with guile 1.8.9, 2.1.x, 2.2.x,
or 3.0.x.

I don't know what changed in guile to cause this problem for this one
release, but looking at the code, I did notice that there's no real
reason why we need to build up this large list.  We could instead,
build a single string, adding the new content to the end of the string
in each pass.  I don't know if this is more or less efficient (I would
guess less efficient), but right now I'd prefer to prioritise
functional over efficient, and supporting a wider set of guile
versions means that CGEN is more likely to "just work" for a users
given setup.

So, in this commit I rewrite the recursive core of gen-enum-decl to
build a single large string instead of a list of small strings.  This
large string is then merged with the prefix and suffix in the
non-recursive outer part of gen-enum-decl.

I tested this regenerating all the opcodes files for all targets, and
there's no difference in any of the output.

ChangeLog:

	* enum.scm (gen-enum-decl): Build a string, rather than a string
	list, within the recursive core of this function.