Update cgen-debugging-stack-start for guile 2+
As we look to support guile 2+ the save-stack function is no longer
supported.  This commit rewrites cgen-debugging-stack-start to avoid
using save-stack.

I'm now using make-stack and display-backtrace instead, which I
believe provides equivalent functionality.  I have also removed the
call to start-stack as I don't think this is needed.  I tried a few
experiments calling (throw ....) from within the cgen core and I
couldn't see any difference with or without the call to start-stack,
so I removed the call.

Finally, the handler currently only prints the backtrace some of the
time (if the exception has enough arguments), I've modified this so we
always print a backtrace if we get an exception.  This certainly
doesn't cause any problems in a "normal" (non-error) run of CGEN, but
maybe this will reveal itself to be a mistake later down the line.

ChangeLog:

	* guile.scm (cgen-debugging-stack-start): Always print the
	backtrace no matter the argument count, use display-backtrace and
	make-stack instead of save-stack and backtrace.  Remove a comment
	that I think no longer applies.  Remove use of start-stack.